% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{addLFC}
\alias{addLFC}
\title{Add Log Fold Change Relative to Reference Group}
\usage{
addLFC(X, parallel = F)
}
\arguments{
\item{X}{A \code{DCSexp} object.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\description{
Note: Add LFC after normalization.
}
