% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{apply.FSexp}
\alias{apply.FSexp}
\title{Apply to FS Experiment}
\usage{
apply.FSexp(X, f, exp.out = T, parallel = F, ...)
}
\arguments{
\item{X}{A \code{FSexp} object.}

\item{f}{Function to apply to each chromosome in \code{FSexp@Data}. Should 
accept objects of class \code{FSchr}.}

\item{exp.out}{Is the output an object of class \code{FSexp}? Default is 
TRUE.}

\item{parallel}{Run in parallel? Must register parallel backend first. 
Default is FALSE.}

\item{...}{Additional arguments for f.}
}
\value{
List resulting from application of function f to the chromosomes in
  \code{FSexp@Data}. If \code{exp.out=T}, the list is formatted as a
  \code{FSchr}.
}
\description{
Apply to FS Experiment
}
