% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{as.data.frame.DCSexp}
\alias{as.data.frame.DCSexp}
\title{as.data.frame.DCSexp}
\usage{
\method{as.data.frame}{DCSexp}(x, row.names, optional, ...)
}
\arguments{
\item{x}{A \code{DCSexp} object.}

\item{row.names}{Unused.}

\item{optional}{Unused.}

\item{...}{Unused.}
}
\description{
as.data.frame.DCSexp
}
