% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{as.data.frame.FSchr}
\alias{as.data.frame.FSchr}
\title{as.data.frame.FSchr}
\usage{
\method{as.data.frame}{FSchr}(x, row.names, optional, ...)
}
\arguments{
\item{x}{An \code{FSchr} object.}

\item{row.names}{Unused.}

\item{optional}{Unused.}

\item{...}{Unused.}
}
\description{
as.data.frame.FSchr
}
