% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{as.data.frame.FSexp}
\alias{as.data.frame.FSexp}
\title{as.data.frame.FSexp}
\usage{
\method{as.data.frame}{FSexp}(x, row.names, optional, ...)
}
\arguments{
\item{x}{An \code{FSexp} object.}

\item{row.names}{Unused.}

\item{optional}{Unused.}

\item{...}{Unused.}
}
\description{
as.data.frame.FSexp
}
