% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{Auxin_rep1}
\alias{Auxin_rep1}
\title{Auxin_rep1}
\format{A list of two sparse matrices of class dtCMatrix \describe{ 
  \item{chr1}{Sample of data from chromosome one.} 
  \item{chr2}{Sample of data from chromosome two.} 
  }}
\usage{
Auxin_rep1
}
\description{
Sample of HiC data from MES cells depleated of CTCF using auxin. Data obtained
from GEO, accession GSE98671, sample GSM2644947.
}
\keyword{datasets}
