% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{B}
\alias{B}
\title{Border Scores}
\format{A \code{FSexp} with two \code{FSchr}s \describe{
    \item{\code{B@Data[["1"]]}}{\code{FSchr} with border scores for chromosome 1.}
    \item{\code{B@Data[["2"]]}}{\code{FSchr} with border scores  for chromosome 2.} 
}}
\usage{
B
}
\description{
An \code{FSexp} containing border scores calculated using the normalized HiC
data in \code{\link{Z}}.
}
\keyword{datasets}
