% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{BorderScores}
\alias{BorderScores}
\title{Experiment-level Border Score Calculation}
\usage{
BorderScores(X, agg = sum, parallel = F)
}
\arguments{
\item{X}{HiC experiment.}

\item{agg}{Function used to aggregate signal within each region.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\value{
Returns a list with one data.frame per group.
}
\description{
Calculates a border scores by comparing the mean pixel intensity in 
triangular windows upstream and downstream of the focus to the mean pixel 
intensity in the intervening region.
}
\details{
minPix refers to the minimum number of pixels observed in one of the
aggregation regions. minProp refers to the minimum proportion of pixels
observed in one of the aggregation regions.
}
