% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{borderWrap}
\alias{borderWrap}
\title{Wrapper Function for Chromosome-level Border Score Calculation}
\usage{
borderWrap(A, agg = sum)
}
\arguments{
\item{A}{Chromosome-level data from HiC experiment.}

\item{agg}{FUNCTION for signal aggregation within each region. Defaults to sum.}
}
\description{
Wrapper Function for Chromosome-level Border Score Calculation
}
