% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{calcLFC}
\alias{calcLFC}
\title{Calculate log2 Fold Change}
\usage{
calcLFC(Counts, g)
}
\arguments{
\item{Counts}{Pixel intensities}

\item{g}{Sample group assignments}
}
\description{
Calculate log2 Fold Change
}
