% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{callPeaks}
\alias{callPeaks}
\title{Call Peaks}
\usage{
callPeaks(foci, b, h, t)
}
\arguments{
\item{foci}{Focus coordinate.}

\item{b}{Border score.}

\item{h}{Window half-width.}

\item{t}{Border score threshold.}
}
\description{
Scans across a chromosome. Calculates the mean and variance of the border
score within the window [focus-h,focus+h]. Identifies foci with a
standardized border score above the threshold.
}
