% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{chrBorderScore}
\alias{chrBorderScore}
\title{Chromosome-level Border Score Calculation}
\usage{
chrBorderScore(G, agg = sum)
}
\arguments{
\item{G}{Chromosome level data structured as (i,j,groupMean)}

\item{agg}{FUNCTION for signal aggregation within each region. Defaults to sum.}
}
\value{
A matrix containing the focus, region summary statistics, and the border
score. Regions (A,B,C) refer to the upstream, downstream, and inter-TAD regions 
respectively. Signal is the aggregated signal in the region. Pixels is the total 
number of pixels in the region, and prop is the proportion of pixels which 
were populated.
}
\description{
For each focus, aggregates signal in the inter-TAD region, and within the
upstream and downstream intra-TAD regions. Calculates a border score as the 
ratio of mean signal in the intra-TAD regions to mean signal in the inter-TAD region.
}
