% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{Control_rep2}
\alias{Control_rep2}
\title{Control_rep2}
\format{A list of two sparse matrices of class dtCMatrix \describe{ 
  \item{chr1}{Sample of data from chromosome one.}
  \item{chr2}{Sample of data from chromosome two.} 
  }}
\usage{
Control_rep2
}
\description{
Sample of HiC data from control MES cells. Data obtained
from GEO, accession GSE98671, sample GSM2644946.
}
\keyword{datasets}
