% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{createDCSexp}
\alias{createDCSexp}
\title{Combine Multiple Samples into a DCS Experiment}
\usage{
createDCSexp(L, smps, groups)
}
\arguments{
\item{L}{List of DCSexp objects corresponding to different samples}

\item{smps}{Sample names, required if not already assigned to component samples}

\item{groups}{Optional group assignments}
}
\value{
An object of class \code{DCSexp}.
}
\description{
Combine Multiple Samples into a DCS Experiment
}
\examples{
# Auxin 
A1 = DCS::Auxin_rep1;
A2 = DCS::Auxin_rep2;
# Control
C1 = DCS::Control_rep1;
C2 = DCS::Control_rep2;
# Combine sparse matrices for individual chromosomes within a sample
Auxin1 = DCS::createDCSsmp(L=A1,chr=c("1","2"),smp="Aux.R1",group=0);
Auxin2 = DCS::createDCSsmp(L=A2,chr=c("1","2"),smp="Aux.R2",group=0);
Ctrl1 = DCS::createDCSsmp(L=C1,chr=c("1","2"),smp="Ctrl.R1",group=1);
Ctrl2 = DCS::createDCSsmp(L=C2,chr=c("1","2"),smp="Ctrl.R2",group=1);
# Combine samples into an experiment
Exp = DCS::createDCSexp(L=list(Auxin1,Auxin2,Ctrl1,Ctrl2));
}
