% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{createDCSsmp}
\alias{createDCSsmp}
\title{Convert List of Sparse Matrices to a DCS Experiment}
\usage{
createDCSsmp(L, chr, smp, group)
}
\arguments{
\item{L}{List of sparse matrices}

\item{chr}{Vector of chromosome names, one per sparse matrix}

\item{smp}{Optional sample name}

\item{group}{Optional group assignment}
}
\value{
An object of class \code{DCSexp}.
}
\description{
Convert List of Sparse Matrices to a DCS Experiment
}
\examples{
# Auxin 
A1 = DCS::Auxin_rep1;
A2 = DCS::Auxin_rep2;
# Combine sparse matrices for individual chromosomes within a sample
Auxin1 = DCS::createDCSsmp(L=A1,chr=c("1","2"),smp="Aux.R1",group=0);
Auxin2 = DCS::createDCSsmp(L=A2,chr=c("1","2"),smp="Aux.R2",group=0);
}
