% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{DCSchr-class}
\alias{DCSchr-class}
\title{DCS Chromosome}
\description{
DCS object for storing chromosome level data.
}
\section{Slots}{

\describe{
\item{\code{chr}}{Chromosome, character.}

\item{\code{Coord}}{Matrix of pixel coordinates.}

\item{\code{Counts}}{Matrix of pixel intensities.}

\item{\code{groups}}{Group assignments, factor.}

\item{\code{samples}}{Sample names, character.}

\item{\code{Stats}}{Statistics available for each pixel, data.frame.}
}}

