% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{methods}
\name{show,DCSchr-method}
\alias{show,DCSchr-method}
\title{show for DCSchr}
\usage{
\S4method{show}{DCSchr}(object)
}
\arguments{
\item{object}{A \code{DCSchr} object.}
}
\description{
show for DCSchr
}
