% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{DCSexp-class}
\alias{DCSexp-class}
\title{DCS Experiment}
\description{
DCS object for storing experiment level data.
}
\section{Slots}{

\describe{
\item{\code{chrs}}{Chromosomes in experiment, character.}

\item{\code{Data}}{List of DCSchrs.}

\item{\code{groups}}{Group assignments, factor.}

\item{\code{samples}}{Sample names, character.}
}}

