% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{methods}
\name{show,DCSexp-method}
\alias{show,DCSexp-method}
\title{show for DCSexp}
\usage{
\S4method{show}{DCSexp}(object)
}
\arguments{
\item{object}{A \code{DCSexp} object.}
}
\description{
show for DCSexp
}
