% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{diagMeans}
\alias{diagMeans}
\title{Diagonal Means}
\usage{
diagMeans(A)
}
\arguments{
\item{A}{DCS chromosome.}
}
\description{
Calculate the mean intensity across samples for each pixel separation
\emph{d}.
}
