% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{diagNormalize}
\alias{diagNormalize}
\title{Normalize HiC Samples}
\usage{
diagNormalize(X, parallel = F)
}
\arguments{
\item{X}{HiC experiment.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\description{
Normalizes a HiC experiment s.t. all samples have the same mean pixel intensity at each 
possible pixel separation \eqn{d\in\{1,\cdots,d_{\max}\}}.
}
