% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{diagScale}
\alias{diagScale}
\title{Rescale Intensities}
\usage{
diagScale(A, S)
}
\arguments{
\item{A}{DCS chromosome.}

\item{S}{Matrix of scale factors}
}
\description{
Rescales the observed HiC intensities along the diagonals of the contact matrix.
}
