% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{dim.DCSchr}
\alias{dim.DCSchr}
\title{dim.DCSchr}
\usage{
\method{dim}{DCSchr}(x)
}
\arguments{
\item{x}{A \code{DCSchr} object.}
}
\value{
Numeric vector, with (number of pixels, number of samples).
}
\description{
dim.DCSchr
}
