% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{dim.DCSexp}
\alias{dim.DCSexp}
\title{dim.DCSexp}
\usage{
\method{dim}{DCSexp}(x)
}
\arguments{
\item{x}{A \code{DCSexp} object.}
}
\value{
A data.frame, with (chromosome, pixels, max pixel separation).
}
\description{
dim.DCSexp
}
