% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{dim.FSchr}
\alias{dim.FSchr}
\title{dim.FSchr}
\usage{
\method{dim}{FSchr}(x)
}
\arguments{
\item{x}{An \code{FSchr} object.}
}
\value{
A numeric vector, with (number of foci).
}
\description{
dim.FSchr
}
