% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{dim.FSexp}
\alias{dim.FSexp}
\title{dim.FSexp}
\usage{
\method{dim}{FSexp}(x)
}
\arguments{
\item{x}{A \code{FSexp} object.}
}
\value{
A data.frame, with (chromosomes, number of foci).
}
\description{
dim.FSexp
}
