% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{expandExp}
\alias{expandExp}
\title{Expand HiC Experiment}
\usage{
expandExp(X)
}
\arguments{
\item{X}{HiC experiment.}
}
\description{
Expands a collapsed HiC experiment into a list with 
one data frame per chromosome.
}
