% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{expApply}
\alias{expApply}
\title{Apply to HiC Experiment}
\usage{
expApply(X, f, cores, ...)
}
\arguments{
\item{X}{HiC experiment.}

\item{f}{Function to apply.}

\item{...}{Additional arguments for f.}
}
\description{
Apply to HiC Experiment
}
