% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{extractData}
\alias{extractData}
\title{Extract Data from DCSchr}
\usage{
extractData(M)
}
\arguments{
\item{M}{A \code{DCSchr} object.}
}
\description{
Extracts the coordinates, counts, and stats from a DCS chromosome
}
