% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{filterLFC}
\alias{filterLFC}
\title{Filter Based on Log Fold Change Relative to Reference Group}
\usage{
filterLFC(X, aT, qT, parallel = F)
}
\arguments{
\item{X}{A \code{DCSexp} object.}

\item{aT}{Absolute LFC threshold}

\item{qT}{Quantile for LFC thresholding}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\description{
Filter Based on Log Fold Change Relative to Reference Group
}
