% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{FSchr-class}
\alias{FSchr-class}
\title{FS Chromosome}
\description{
Object for storing focal statistics (FS) at the chromosome level.
}
\section{Slots}{

\describe{
\item{\code{chr}}{Chromosome, character.}

\item{\code{foci}}{Foci, or loci along the diagonal of the contact matrix, for which statistics are available.}

\item{\code{groups}}{Group assignments, factor.}

\item{\code{Stats}}{List of statistics, one per group.}
}}

