% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{methods}
\name{show,FSchr-method}
\alias{show,FSchr-method}
\title{show for FSchr}
\usage{
\S4method{show}{FSchr}(object)
}
\arguments{
\item{object}{An \code{FSchr} object.}
}
\description{
show for FSchr
}
