% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{class}
\name{FSexp-class}
\alias{FSexp-class}
\title{FS Experiment}
\description{
Object for storing focal statistics (FS) at the experiment level.
}
\section{Slots}{

\describe{
\item{\code{chrs}}{Chromosomes, character.}

\item{\code{groups}}{Group assignments, factor.}

\item{\code{Data}}{List of FSchrs.}
}}

