% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\docType{methods}
\name{show,FSexp-method}
\alias{show,FSexp-method}
\title{show for FSexp}
\usage{
\S4method{show}{FSexp}(object)
}
\arguments{
\item{object}{An \code{FSexp} object.}
}
\description{
show for FSexp
}
