% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gas_gc_llCpp}
\alias{gas_gc_llCpp}
\title{Log-likelihood of the GAS Gaussian copula model}
\usage{
gas_gc_llCpp(vParams, iT, mData)
}
\description{
Returns the scaled negative log-likelihood of the GAS Gaussian copula model for a given
set of parameters and data matrix
}

