% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{genDiag}
\alias{genDiag}
\title{Coordinates of Matrix Diagonals}
\usage{
genDiag(a, b, d)
}
\arguments{
\item{a}{Stat of sequence}

\item{b}{End of sequence}

\item{d}{Maximum pixel separation}
}
\description{
Coordinates of Matrix Diagonals
}
