% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{getBorderScores}
\alias{getBorderScores}
\title{Experiment-level Border Score Calculation}
\usage{
getBorderScores(X, s, g, agg = sum, cores = 1)
}
\arguments{
\item{X}{HiC experiment.}

\item{s}{Sample names.}

\item{g}{Sample group assignments.}

\item{agg}{agg FUNCTION for signal aggregation within each region. Defaults
to sum.}

\item{cores}{Cores to use if running in parallel.}
}
\value{
Returns a list with one data.frame per group.
}
\description{
Calculates a border scores by comparing the mean pixel intensity in 
triangular windows upstream and downstream of the focus to the mean pixel 
intensity in the intervening region.
}
\details{
minPix refers to the minimum number of pixels observed in one of the
aggregation regions. minProp refers to the minimum proportion of pixels
observed in one of the aggregation regions.
}
