% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{ggHiC}
\alias{ggHiC}
\title{ggplot HiC Heatmap}
\usage{
ggHiC(df, Foci, h.u, h.d, l.u)
}
\arguments{
\item{df}{Plotting frame}

\item{Foci}{Sequence of foci plotted}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Length of upstream flank in pixels}
}
\description{
Contains the ggplot commands for creating the HiC heatmap.
}
