% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{head.DCSchr}
\alias{head.DCSchr}
\title{head.DCSchr}
\usage{
\method{head}{DCSchr}(x, ..., n = 6)
}
\arguments{
\item{x}{A \code{DCSchr} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}
}
\description{
head.DCSchr
}
