% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{head.DCSexp}
\alias{head.DCSexp}
\title{head.DCSexp}
\usage{
\method{head}{DCSexp}(x, ..., n = 6)
}
\arguments{
\item{x}{A \code{DCSexp} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}
}
\description{
head.DCSexp
}
