% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{head.FSchr}
\alias{head.FSchr}
\title{head.FSchr}
\usage{
\method{head}{FSchr}(x, ..., n = 6)
}
\arguments{
\item{x}{An \code{FSchr} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}
}
\description{
head.FSchr
}
