% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{head.FSexp}
\alias{head.FSexp}
\title{head.FSexp}
\usage{
\method{head}{FSexp}(x, ..., n = 6)
}
\arguments{
\item{x}{An \code{FSexp} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}
}
\description{
head.FSexp
}
