% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{impute}
\alias{impute}
\title{Impute Missing Pixels}
\usage{
impute(Counts, Coords)
}
\arguments{
\item{Counts}{Matrix of pixel intensities}

\item{Coords}{Matrix of pixel coordinates}
}
\description{
Impute Missing Pixels
}
