% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{merge.DCSchr}
\alias{merge.DCSchr}
\title{merge.DCSchr}
\usage{
\method{merge}{DCSchr}(x, y, ...)
}
\arguments{
\item{x}{A \code{DCSchr}.}

\item{y}{Another \code{DCSchr}.}

\item{...}{Unused.}
}
\value{
A \code{DCSchr}.
}
\description{
merge.DCSchr
}
