% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{mExclusion}
\alias{mExclusion}
\title{Matrix Exclusion}
\usage{
mExclusion(A, B)
}
\arguments{
\item{A}{First matrix.}

\item{B}{Second matrix.}
}
\description{
Find all rows in matrix A NOT in matrix B
}
