% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{missedCoord}
\alias{missedCoord}
\title{Coordinates of Missing Pixels}
\usage{
missedCoord(Coords)
}
\arguments{
\item{Coords}{Matrix of pixel coordinates}
}
\description{
Enumerates coordinates of pixels NOT present in the data.
}
