% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Pixels}
\alias{Pixels}
\title{Heatmap Pixel Coords}
\usage{
Pixels(s)
}
\arguments{
\item{s}{Sequence of foci.}
}
\value{
Data.frame containing pixel coordinates (i,j) and a unique identifier in
  the form (1+i-min(i),1+j-min(i)).
}
\description{
Given a sequence of foci, enumerates coordinates of all pixels included in 
the heatmap.
}
