% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotHicBasic}
\alias{plotHicBasic}
\title{Basic HiC Plot}
\usage{
plotHicBasic(A, f, h.u, h.d, l.u, l.d)
}
\arguments{
\item{A}{Chromosome-level data as (i,j,stat)}

\item{f}{Focus.}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Lenght of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}
}
\description{
Creates a HiC heatmap from data of the form (i,j,stat), where stat is a continuous
statistic.
}
