% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotHicGrouped}
\alias{plotHicGrouped}
\title{HiC Plot by Group}
\usage{
plotHicGrouped(A, f, h.u, h.d, l.u, l.d, smooth = F)
}
\arguments{
\item{A}{Chromosome-level data as  (i,j, group1 stat, group2 stat, etc.)}

\item{f}{Focus.}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Lenght of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}

\item{Smooth}{Should pixel intensity be smoothed?}
}
\description{
Creates a HiC heatmap faceted by group.
}
