% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotHicSignal}
\alias{plotHicSignal}
\title{Plot Log HiC Signal}
\usage{
plotHicSignal(In, f, chr, h.u, h.d, l.u, l.d, smooth = F)
}
\arguments{
\item{In}{A \code{DCSexp} or \code{DCSchr} object.}

\item{f}{Plot focus, numeric.}

\item{chr}{Chromosome, if providing a \code{DCSexp}.}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Lenght of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}

\item{smooth}{Smooth signal against pixel separation \eqn{d}?}
}
\value{
List containing 1. the group facetted heatmaps, 2. a list of the
  component plots, and 3. a data.frame containing the minimum and maximal 
  signal within the plotting region for each group.
}
\description{
Creates a heatmap displaying the log of the HiC signal in each pixel. The 
plots are facetted by group. For groups with multiple samples, the mean
signal is taken prior to the logarithm.
}
