% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotHicSmoothed}
\alias{plotHicSmoothed}
\title{HiC Plot with Domain Smoothing}
\usage{
plotHicSmoothed(A, f, h.u, h.d, l.u, l.d)
}
\arguments{
\item{A}{Chromosome-level data as (i,j,stat)}

\item{f}{Focus.}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Lenght of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}
}
\description{
HiC Plot with Domain Smoothing
}
