% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotHicStat}
\alias{plotHicStat}
\title{Plot HiC Statistic}
\usage{
plotHicStat(In, stat, f, chr, h.u, h.d, l.u = 0, l.d = 0, smooth = F)
}
\arguments{
\item{In}{A \code{DCSexp} or \code{DCSchr} object.}

\item{stat}{Name of statistic to plot. Should correspond with a column of
\code{DCSchr@Stats}.}

\item{f}{Plot focus, numeric.}

\item{chr}{Chromosome, if providing a \code{DCSexp}.}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Lenght of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}

\item{smooth}{Smooth signal against pixel separation \eqn{d}?}
}
\value{
List containing 1. the ggplot, 2. the minimal signal observed in the 
  plotting region, and 3. the maximal signal in the plotting region.
}
\description{
Creates a heatmap displaying the value of a HiC statistic within each pixel.
}
