% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{plotNormCurves}
\alias{plotNormCurves}
\title{Plot HiC Normalization Curves}
\usage{
plotNormCurves(X, Z, parallel = F)
}
\arguments{
\item{X}{Initial HiC experiment.}

\item{Z}{HiC Experiment after Normalization.}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\description{
Plots mean signal intensity by pixel separation before and after
HiC normalization.
}
