% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{polyCoord}
\alias{polyCoord}
\title{Enumerate polygon vertices}
\usage{
polyCoord(h.u, h.d, l.u, l.d)
}
\arguments{
\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Length of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}
}
\value{
List containing 1. the polygon centers, and 2. the polygon vertices. 
  \code{Centers} is a data.frame containing (x,y) coordinates of the pixel
  centers and the pixel id. \code{Vertices} is a data.frame containing (x,y)
  coordinates of the pixel vertices, the pixel id, and the order in which
  vertices are connected.
}
\description{
Given the base height and flank, enumerates the vertices of pixels that are
connected to form the heatmap.
}
