% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{polyX}
\alias{polyX}
\title{X-Coord of Pixel Vertices}
\usage{
polyX(r)
}
\arguments{
\item{r}{Pixel coords (i,j);}
}
\value{
Numeric vector of x-coordinates for heatmap.
}
\description{
Starting from the left-most vertex and proceeding counter-clockwise,
enumerates the x-coordinates of a pixel's vertices.
}
