% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{preProcess}
\alias{preProcess}
\title{Pre-process HiC Experiment}
\usage{
preProcess(X, miss = T, zero = T, parallel = F)
}
\arguments{
\item{X}{HiC experiment.}

\item{miss}{Impute missing pixels to zeros?}

\item{zero}{Remove pixels corresponding to empty foci?}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\description{
Pre-process HiC Experiment
}
