% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{print.DCSchr}
\alias{print.DCSchr}
\title{print.DCSchr}
\usage{
\method{print}{DCSchr}(x, ..., n = 6)
}
\arguments{
\item{x}{A \code{DCSchr} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}
}
\description{
print.DCSchr
}
