% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{print.DCSexp}
\alias{print.DCSexp}
\title{print.DCSexp}
\usage{
\method{print}{DCSexp}(x, ..., n = 6, chrs)
}
\arguments{
\item{x}{A \code{DCSexp} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}

\item{chrs}{Chromosomes to print, character.}
}
\description{
print.DCSexp
}
