% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{print.FSchr}
\alias{print.FSchr}
\title{print.FSchr}
\usage{
\method{print}{FSchr}(x, ..., n = 6)
}
\arguments{
\item{x}{An \code{FSchr} object.}

\item{...}{Unused.}

\item{n}{Number of observations to print.}
}
\description{
print.FSchr
}
