% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{print.FSexp}
\alias{print.FSexp}
\title{print.FSexp}
\usage{
\method{print}{FSexp}(x, ..., chrs, n = 6)
}
\arguments{
\item{x}{An \code{FSexp} object.}

\item{...}{Unused.}

\item{chrs}{Chromosomes to print, character.}

\item{n}{Number of observations to print.}
}
\description{
print.FSexp
}
