% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{pruneEmpty}
\alias{pruneEmpty}
\title{Remove Pixels Corresponding to Empty Foci}
\usage{
pruneEmpty(A)
}
\arguments{
\item{A}{Chromosome level data}
}
\description{
Remove Pixels Corresponding to Empty Foci
}
