% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{scaleFactors}
\alias{scaleFactors}
\title{Scale Factor Estimation}
\usage{
scaleFactors(X, parallel = F)
}
\arguments{
\item{X}{HiC experiment}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\description{
Function to calculate separation based scale factors. In particular, the scale factor
for the \emph{i}th sample is calcuate as the mean across samples at separation \emph{d}, \eqn{m(d)},
divided by the mean in sample \emph{i} at separation \emph{d}, \eqn{m_{i}(d)}.
}
