% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{sHicDatum}
\alias{sHicDatum}
\title{sHiCDatum}
\format{A sparseHiCdatum \describe{
    \item{sHicDatum@sampleName}{Here, auxin_rep1.}
    \item{sHicDatum@resolutionNamedList}{List of sparse contact matrices, indexed by resolution.}
    \item{sHicDatum@metaData}{Reference genome.}
}}
\usage{
sHicDatum
}
\description{
Sample of HiC data from MES cells depleated of CTCF using auxin. Data obtained
from GEO, accession GSE98671, sample GSM2644947. These data are in sparse HiC format.
}
\keyword{datasets}
