% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Smooth}
\alias{Smooth}
\title{Loess Smooth against Separation}
\usage{
Smooth(X)
}
\arguments{
\item{X}{Matrix of the form (i,j,Signal)}
}
\value{
Smoothed matrix.
}
\description{
Loess Smooth against Separation
}
