% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{smoothHicPlot}
\alias{smoothHicPlot}
\title{HiC Plot with Domain Smoothing}
\usage{
smoothHicPlot(B, f, h.u, h.d, l.u, l.d)
}
\arguments{
\item{B}{Data.table in the format (i,j,Signal)}

\item{f}{Plot focus, numeric.}

\item{h.u}{Length of upstream domain in pixels}

\item{h.d}{Length of downstream domain in pixels}

\item{l.u}{Lenght of upstream flank in pixels}

\item{l.d}{Length of downstream flank in pixels}
}
\value{
List containing 1. the ggplot, 2. the minimal signal observed in the 
  plotting region, and 3. the maximal signal in the plotting region.
}
\description{
Creates a HiC heatmap displaying the fitted value of the test statistic smoothed
within domains by pixel separation.
}
