% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{sparseMat2DCS}
\alias{sparseMat2DCS}
\title{Sparse Matrix to DCS Chromosome}
\usage{
sparseMat2DCS(A, chr, smp, group)
}
\arguments{
\item{A}{Sparse matrix of class dtCMatrix}

\item{chr}{Chromosome}

\item{smp}{Sample name}

\item{group}{Group name}
}
\value{
A contact matrix of class \code{DCSchr}.
}
\description{
Sparse Matrix to DCS Chromosome
}
