% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{subFlank}
\alias{subFlank}
\title{Subset Flank Pixels}
\usage{
subFlank(R, S)
}
\arguments{
\item{R}{All pixels in region}

\item{S}{Non-flank pixels in region}
}
\description{
Given 1. coordinates for all pixels in a region, and 2. coordinates for non-flank
pixels in the region, subsets the coordinates of flank pixels.
}
