% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{subset.DCSexp}
\alias{subset.DCSexp}
\title{subset.DCSexp}
\usage{
\method{subset}{DCSexp}(x, ..., chrs, dmax)
}
\arguments{
\item{x}{A \code{DCSexp} object.}

\item{...}{Unused.}

\item{chrs}{Chromosomes to retain, character.}

\item{dmax}{Maximal pixel separation, numeric.}
}
\description{
Allows subsetting of a DCSexp by chromosome and/or maximal pixel separation.
}
\examples{
# Subset pixels on chromosome 1 with maximal separation 10
subset(x=DCS::Z,chrs="1",dmax=10);
}
