% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{subset.FSchr}
\alias{subset.FSchr}
\title{subset.FSchr}
\usage{
\method{subset}{FSchr}(x, ..., foci, groups)
}
\arguments{
\item{x}{An \code{FSchr} object.}

\item{...}{Unused}

\item{foci}{Foci to reain, logical vector.}

\item{groups}{Groups to retain, character or factor.}
}
\description{
Allows subsetting an FSchr using a logical vector and/or the group assignment.
}
