% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{subset.FSexp}
\alias{subset.FSexp}
\title{subset.FSexp}
\usage{
\method{subset}{FSexp}(x, ..., chrs, groups)
}
\arguments{
\item{x}{FSexp.}

\item{...}{Unused.}

\item{chrs}{Chromosomes to retain, character.}

\item{groups}{Groups to retain, character or factor.}
}
\description{
Allows subsetting an FSexp by chromosome and/or group.
}
