% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefinitions.R
\name{summary.DCSexp}
\alias{summary.DCSexp}
\title{summary.DCSexp}
\usage{
summary.DCSexp(object, ...)
}
\arguments{
\item{object}{A \code{DCSexp} object.}

\item{...}{Unused.}
}
\value{
A list containing 1. the dimensions of each chromosome in the 
  experiment, and 2. the group assignments.
}
\description{
summary.DCSexp
}
