% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{thinPeaks}
\alias{thinPeaks}
\title{Thin Peaks}
\usage{
thinPeaks(p, b, w)
}
\arguments{
\item{p}{Peak coordinates}

\item{b}{Peak border scores}

\item{w}{Minimum separation}
}
\description{
If multiple candidate peaks are within the minimum separation distance,
evaluates the border score at each, and retains the representative peak with
the highest border score.
}
