% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{triCoord}
\alias{triCoord}
\title{Purpose : Functions to calculate a border score for each focus in the genome.
Border scores compare the signal intensity in a putative inter-TAD region with
the mean border score in the flanking regions. 
Enumerate Inter-Tad Region}
\usage{
triCoord(f, d, l, u)
}
\arguments{
\item{f}{Focus.}

\item{d}{Maximum pixel separation.}

\item{l}{Lower bound on pixel indices.}

\item{u}{Upper bound on pixel indices.}
}
\description{
Enumerates pixels in an upper triangular window above a focus. Maximal 
pixel separation d corresond to the base-width of the triangle in pixels. 
Assuming the pixels are indexed (i,j) with i <= j, the lower bound l = min(i), 
while the upper bound u = max(j).
}
