% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{triCoordLeft}
\alias{triCoordLeft}
\title{Enumerate Upstream Intra-TAD Region}
\usage{
triCoordLeft(f, d, l, u)
}
\arguments{
\item{f}{Focus.}

\item{d}{Maximum pixel separation.}

\item{l}{Lower bound on pixel indices.}

\item{u}{Upper bound on pixel indices.}
}
\description{
Enumerates pixels in an upper triangular window upstream of a focus. Maximal 
pixel separation d corresond to the base-width of the triangle in pixels. 
Assuming the pixels are indexed (i,j) with i <= j, the lower bound l = min(i), 
while the upper bound u = max(j).
}
