% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{wrapLfcFilter}
\alias{wrapLfcFilter}
\title{Filter on Log Fold Change}
\usage{
wrapLfcFilter(A, aT, qT)
}
\arguments{
\item{A}{Chromosome level data}

\item{aT}{Absolute threshold}

\item{qT}{Quantile threshold}
}
\description{
Filter on Log Fold Change
}
