% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{wrapPP}
\alias{wrapPP}
\title{Wrapper Function for Pre-processing}
\usage{
wrapPP(A, miss = T, zero = T)
}
\arguments{
\item{A}{Chromosome level data}

\item{miss}{Impute missing pixels to zeros?}

\item{zero}{Remove pixels corresponding to empty foci?}
}
\description{
Wrapper Function for Pre-processing
}
