% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{Z}
\alias{Z}
\title{Normalized HiC Experiment}
\format{A \code{DCSexp} with two \code{DCSchr}s \describe{
    \item{\code{Z@Data[["1"]]}}{\code{DCSchr} with HiC data for chromosome 1.}
    \item{\code{Z@Data[["2"]]}}{\code{DCSchr} with HiC data for chromosome 2.}
}}
\usage{
Z
}
\description{
A \code{DCSexp} containing pre-processed and normalized HiC data for control
(\code{\link{Control_rep1}}, \code{\link{Control_rep2}}) and auxin treated
(\code{\link{Auxin_rep1}}, \code{\link{Auxin_rep2}}) MES cells.
}
\keyword{datasets}
