% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BorderScore.R
\name{borderWrap}
\alias{borderWrap}
\title{Wrapper Function for Chromosome-level Border Score Calculation}
\usage{
borderWrap(W, s, g, agg)
}
\arguments{
\item{W}{Chromosome-level data from HiC experiment.}

\item{s}{Sample names.}

\item{g}{Group assignments.}

\item{agg}{FUNCTION for signal aggregation within each region. Defaults to sum.}
}
\description{
Wrapper Function for Chromosome-level Border Score Calculation
}
