% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{polyY}
\alias{polyY}
\title{Y-Coord of Pixel Vertices}
\usage{
polyY(r)
}
\arguments{
\item{r}{Pixel coords (i,j);}
}
\description{
Starting from the left-most vertex and proceeding clockwise, enumerates the
y-coordinates of a pixel's vertices.
}
