% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Smooth}
\alias{Smooth}
\title{Smooth Signal}
\usage{
Smooth(X)
}
\arguments{
\item{X}{Data table of the form (i,j,Signal)}
}
\value{
Data table including the coordinates, original signal, pixel
  separation, and fitted values
}
\description{
Smooths the signal against pixel separation. If there are
fewer than 10 observations, a simple mean is taken. If between 10 and 20
observations, linear loess is applied. If greater than 20 observations,
quadratic loess is applied.
}
